+++ app/Http/Controllers/Api/Dashboard/ProductImageController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Storage;
 use App\Interfaces\ImageServiceInterface;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class ProductImageController extends Controller
@@ -22,11 +23,14 @@ class ProductImageController extends Controller
 	public function store(Request $request, $id)
 	{
 		try {
+			$this->authorize('update', Product::class);
 			$storagePath = '/products_images/' . $id;
 
 			$path = $this->imageService->storeImage($request->file('image'), $storagePath, 'products', 'image', $id);
 
 			return response()->json(['imagePath' => $path], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
 		} catch (\Exception $e) {
 			return response()->json(['message' => $e->getTrace()], 500);
 		}
@@ -35,11 +39,14 @@ class ProductImageController extends Controller
 	public function destroy(Request $request, $id)
 	{
 		try {
+			$this->authorize('update', Product::class);
 			DB::beginTransaction();
 			$this->imageService->deleteImage('products', 'image', $id, $request->imagePath, '/storage/products_images/placeholder.jpg');
 
 			DB::commit();
 			return response()->json(['message' => 'Image removed'], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
 		} catch (\Exception $e) {
 			DB::rollBack();
 			debug($e);
